+++ junit/kubernetes-junit-jupiter/src/main/java/io/fabric8/junit/jupiter/NamespaceExtension.java
@@ -15,10 +15,10 @@
  */
 package io.fabric8.junit.jupiter;
 
+import io.fabric8.junit.jupiter.api.KubernetesTest;
 import io.fabric8.kubernetes.api.model.Namespace;
 import io.fabric8.kubernetes.api.model.NamespaceBuilder;
 import io.fabric8.kubernetes.api.model.ObjectReference;
-import io.fabric8.kubernetes.client.Client;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClientBuilder;
 import io.fabric8.kubernetes.client.NamespacedKubernetesClient;
@@ -29,59 +29,44 @@ import org.junit.jupiter.api.extension.ExtensionContext;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
-import java.util.function.Predicate;
-import java.util.stream.Stream;
 
-public class KubernetesNamespacedTestExtension implements BeforeAllCallback, BeforeEachCallback, AfterAllCallback {
+public class NamespaceExtension implements HasKubernetesClient, BeforeAllCallback, BeforeEachCallback, AfterAllCallback {
 
   @Override
   public void beforeAll(ExtensionContext context) throws Exception {
     final KubernetesClient client = new KubernetesClientBuilder().build();
-    getStore(context).put(Namespace.class, initNamespace(client));
-    getStore(context).put(KubernetesClient.class,
-        client.adapt(NamespacedKubernetesClient.class).inNamespace(getNamespace(context).getMetadata().getName()));
-    for (Field field : extractFields(context, KubernetesClient.class, f -> Modifier.isStatic(f.getModifiers()))) {
-      setFieldValue(field, null, getClient(context).adapt((Class<Client>) field.getType()));
+    getStore(context).put(KubernetesClient.class, client);
+    if (shouldCreateNamespace(context)) {
+      getStore(context).put(Namespace.class, initNamespace(client));
+      getStore(context).put(KubernetesClient.class,
+          client.adapt(NamespacedKubernetesClient.class).inNamespace(getKubernetesNamespace(context).getMetadata().getName()));
     }
     for (Field field : extractFields(context, Namespace.class, f -> Modifier.isStatic(f.getModifiers()))) {
-      setFieldValue(field, null, getNamespace(context));
+      setFieldValue(field, null, getKubernetesNamespace(context));
     }
   }
 
   @Override
   public void beforeEach(ExtensionContext context) throws Exception {
-    for (Field field : extractFields(context, KubernetesClient.class, f -> !Modifier.isStatic(f.getModifiers()))) {
-      setFieldValue(field, context.getRequiredTestInstance(), getClient(context).adapt((Class<Client>) field.getType()));
-    }
     for (Field field : extractFields(context, Namespace.class, f -> !Modifier.isStatic(f.getModifiers()))) {
-      setFieldValue(field, context.getRequiredTestInstance(), getNamespace(context));
+      setFieldValue(field, context.getRequiredTestInstance(), getKubernetesNamespace(context));
     }
   }
 
   @Override
   public void afterAll(ExtensionContext context) {
     final KubernetesClient client = getClient(context);
-    client.resource(getNamespace(context)).withGracePeriod(0L).delete();
-    client.close();
-  }
-
-  static KubernetesClient getClient(ExtensionContext context) {
-    final KubernetesClient client = getStore(context).get(KubernetesClient.class, KubernetesClient.class);
-    if (client == null) {
-      throw new IllegalStateException("No KubernetesClient found");
+    if (shouldCreateNamespace(context)) {
+      client.resource(getKubernetesNamespace(context)).withGracePeriod(0L).delete();
     }
-    return client;
-  }
-
-  private static ExtensionContext.Store getStore(ExtensionContext context) {
-    ExtensionContext.Namespace namespace = ExtensionContext.Namespace.create(KubernetesNamespacedTestExtension.class,
-        context.getRequiredTestClass());
-    return context.getRoot().getStore(namespace);
+    // Note that the ThreadPoolExecutor in OkHttp's RealConnectionPool is shared amongst all the OkHttp client
+    // instances. This means that closing one OkHttp client instance effectively closes all the others.
+    // In order to be able to use this safely, we should transition to one of the other HttpClient implementations
+    client.close();
   }
 
   /**
@@ -120,31 +105,16 @@ public class KubernetesNamespacedTestExtension implements BeforeAllCallback, Bef
     return namespace;
   }
 
-  private static Namespace getNamespace(ExtensionContext context) {
+  private boolean shouldCreateNamespace(ExtensionContext context) {
+    final KubernetesTest annotation = findAnnotation(context.getRequiredTestClass(), KubernetesTest.class);
+    return annotation == null || annotation.createEphemeralNamespace();
+  }
+
+  private Namespace getKubernetesNamespace(ExtensionContext context) {
     final Namespace namespace = getStore(context).get(Namespace.class, Namespace.class);
     if (namespace == null) {
       throw new IllegalStateException("No Kubernetes Namespace found");
     }
     return namespace;
-  }
-
-  private static Field[] extractFields(ExtensionContext context, Class<?> clazz, Predicate<Field>... predicates) {
-    final Class<?> testClass = context.getTestClass().orElse(null);
-    if (testClass != null) {
-      Stream<Field> fieldStream = Arrays.stream(testClass.getDeclaredFields())
-          .filter(f -> clazz.isAssignableFrom(f.getType()));
-      for (Predicate<Field> p : predicates) {
-        fieldStream = fieldStream.filter(p);
-      }
-      return fieldStream.toArray(Field[]::new);
-    }
-    return new Field[0];
-  }
-
-  private static void setFieldValue(Field field, Object entity, Object value) throws IllegalAccessException {
-    final boolean isAccessible = field.isAccessible();
-    field.setAccessible(true);
-    field.set(entity, value);
-    field.setAccessible(isAccessible);
   }
 }