+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/PortForwarderWebsocketListener.java
@@ -15,7 +15,9 @@
  */
 package io.fabric8.kubernetes.client.dsl.internal;
 
+import io.fabric8.kubernetes.client.KubernetesClientException;
 import io.fabric8.kubernetes.client.http.WebSocket;
+import io.fabric8.kubernetes.client.utils.Utils;
 import io.fabric8.kubernetes.client.utils.internal.SerialExecutor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -46,8 +48,6 @@ public class PortForwarderWebsocketListener implements WebSocket.Listener {
 
   private final AtomicBoolean alive = new AtomicBoolean(true);
 
-  private final AtomicBoolean errorOccurred = new AtomicBoolean(false);
-
   final Collection<Throwable> clientThrowables = new CopyOnWriteArrayList<>();
 
   final Collection<Throwable> serverThrowables = new CopyOnWriteArrayList<>();
@@ -75,9 +75,9 @@ public class PortForwarderWebsocketListener implements WebSocket.Listener {
           if (e instanceof InterruptedException) {
             Thread.currentThread().interrupt();
           }
+          logger.debug("Error while writing client data");
           if (alive.get()) {
             clientThrowables.add(e);
-            logger.error("Error while writing client data");
             closeBothWays(webSocket, 1001, "Client error");
           }
         }
@@ -101,21 +101,27 @@ public class PortForwarderWebsocketListener implements WebSocket.Listener {
     }
 
     if (!buffer.hasRemaining()) {
-      errorOccurred.set(true);
-      logger.error("Received an empty message");
+      KubernetesClientException e = new KubernetesClientException("Received an empty message");
+      serverThrowables.add(e);
+      logger.debug("Protocol error", e);
       closeBothWays(webSocket, 1002, PROTOCOL_ERROR);
       return;
     }
 
     byte channel = buffer.get();
     if (channel < 0 || channel > 1) {
-      errorOccurred.set(true);
-      logger.error("Received a wrong channel from the remote socket: {}", channel);
+      KubernetesClientException e = new KubernetesClientException(
+          String.format("Received a wrong channel from the remote socket: %s", channel));
+      serverThrowables.add(e);
+      logger.debug("Protocol error", e);
       closeBothWays(webSocket, 1002, PROTOCOL_ERROR);
     } else if (channel == 1) {
       // Error channel
-      errorOccurred.set(true);
-      logger.error("Received an error from the remote socket");
+      // TODO: read the error
+      KubernetesClientException e = new KubernetesClientException(
+          String.format("Received an error from the remote socket"));
+      serverThrowables.add(e);
+      logger.debug("Server error", e);
       closeForwarder();
     } else {
       // Data
@@ -136,7 +142,7 @@ public class PortForwarderWebsocketListener implements WebSocket.Listener {
             }
             if (alive.get()) {
               clientThrowables.add(e);
-              logger.error("Error while forwarding data to the client", e);
+              logger.debug("Error while forwarding data to the client", e);
               closeBothWays(webSocket, 1002, PROTOCOL_ERROR);
             }
           }
@@ -155,10 +161,9 @@ public class PortForwarderWebsocketListener implements WebSocket.Listener {
 
   @Override
   public void onError(WebSocket webSocket, Throwable t) {
-    logger.debug("{}: onFailure", LOG_PREFIX);
+    logger.debug("{}: Throwable received from websocket", LOG_PREFIX, t);
     if (alive.get()) {
       serverThrowables.add(t);
-      logger.error("{}: Throwable received from websocket", LOG_PREFIX, t);
       closeForwarder();
     }
   }
@@ -168,7 +173,7 @@ public class PortForwarderWebsocketListener implements WebSocket.Listener {
   }
 
   boolean errorOccurred() {
-    return errorOccurred.get() || !clientThrowables.isEmpty() || !serverThrowables.isEmpty();
+    return !clientThrowables.isEmpty() || !serverThrowables.isEmpty();
   }
 
   Collection<Throwable> getClientThrowables() {
@@ -179,14 +184,14 @@ public class PortForwarderWebsocketListener implements WebSocket.Listener {
     return serverThrowables;
   }
 
-  private void closeBothWays(WebSocket webSocket, int code, String message) {
+  void closeBothWays(WebSocket webSocket, int code, String message) {
     logger.debug("{}: Closing with code {} and reason: {}", LOG_PREFIX, code, message);
     alive.set(false);
     try {
       webSocket.sendClose(code, message);
     } catch (Exception e) {
       serverThrowables.add(e);
-      logger.error("Error while closing the websocket", e);
+      logger.debug("Error while closing the websocket", e);
     }
     closeForwarder();
   }
@@ -194,18 +199,10 @@ public class PortForwarderWebsocketListener implements WebSocket.Listener {
   private void closeForwarder() {
     alive.set(false);
     if (in != null) {
-      try {
-        in.close();
-      } catch (IOException e) {
-        logger.error("{}: Error while closing the client input channel", LOG_PREFIX, e);
-      }
+      Utils.closeQuietly(in);
     }
     if (out != null && out != in) {
-      try {
-        out.close();
-      } catch (IOException e) {
-        logger.error("{}: Error while closing the client output channel", LOG_PREFIX, e);
-      }
+      Utils.closeQuietly(out);
     }
     pumperService.shutdownNow();
     serialExecutor.shutdownNow();
@@ -228,4 +225,5 @@ public class PortForwarderWebsocketListener implements WebSocket.Listener {
       }
     } while (isAlive.getAsBoolean() && read >= 0);
   }
+
 }