+++ kubernetes-itests/src/test/java/io/fabric8/kubernetes/GenericClusterScopedCustomResourceIT.java
@@ -51,8 +51,9 @@ class GenericClusterScopedCustomResourceIT {
     client.apiextensions().v1().customResourceDefinitions()
         .withName("satellites.demos.fabric8.io")
         .waitUntilCondition(
-            c -> c.getStatus() != null && c.getStatus().getConditions() != null && c.getStatus().getConditions().stream()
-                .anyMatch(crdc -> crdc.getType().equals("Established") && crdc.getStatus().equals("True")),
+            c -> c != null && c.getStatus() != null && c.getStatus().getConditions() != null
+                && c.getStatus().getConditions().stream()
+                    .anyMatch(crdc -> crdc.getType().equals("Established") && crdc.getStatus().equals("True")),
             10L, TimeUnit.SECONDS);
   }